refactor: use common SpriteList component in default and map editor
